% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dDynOcc.R
\name{dDynOcc}
\alias{dDynOcc}
\alias{rOcc}
\alias{rOcc_s}
\alias{dOcc_v}
\alias{rOcc_v}
\title{Dynamic occupancy distribution for use in NIMBLE models}
\usage{
dDynOcc()
}
\description{
Dynamic occupancy distribution for use in NIMBLE models
}
