% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dHMM.R
\name{dHMM}
\alias{dHMM}
\alias{dHMMo}
\alias{rHMM}
\alias{rHMMo}
\title{Hidden Markov Model distribution for use in NIMBLE models}
\usage{
dHMM(x, init, probObs, probTrans, len = 0, log = 0)

dHMMo(x, init, probObs, probTrans, len = 0, log = 0)

rHMM(n, init, probObs, probTrans, len = 0)

rHMMo(n, init, probObs, probTrans, len = 0)
}
\arguments{
\item{x}{vector of observation classes, one of which could be defined as "not observed".}

\item{init}{vector of initial state probabilities}

\item{probObs}{time-independent matrix of observation probabilities.
Dimension of \code{probObs} is (number of possible observation classes) x (number of possible system states)}

\item{probTrans}{time-independent matrix of system state-transition probabilities.
Dimension of \code{probTrans} is (number of possible system states) x (number of possible system states)}

\item{len}{length of observations (needed for rDHMM)}

\item{log}{TRUE or 1 to return log probability. FALSE or 0 to return probability.}

\item{n}{length of random sequence}
}
\description{
\code{dHMM} and \code{dHMMo} provide hidden Markov model distributions for NIMBLE models.
"Dynamic" here means that the matrix of state transition probabilities is indexed by time.  The
\code{dHMMo} version additionally allows observation probabilities to be indexed by time.
Compared to writing NIMBLE models with discrete latent states, use of these DHMM distributions allows
one to directly integrate over such discrete latent states and hence leave them out of the NIMBLE
model code.
}
\details{
These nimbleFunctions provide distributions that can be used in code (via \link{nimbleCode})
for \link{nimbleModel}.

These are written in the format of user-defined distributions to extend NIMBLE's
use of the BUGS model language.  More information about writing user-defined distributions can be found
in the NIMBLE User Manual at \code{https://r-nimble.org}.

The first argument to a "d" function is always named \code{x} and is given on the
left-hand side of a (stochastic) model declaration in the BUGS model language (used by NIMBLE).
When using these distributions in a NIMBLE model, the user
should not provide the \code{log} argument.  (It is always set to \code{TRUE} when used
in a NIMBLE model.)

For example, in a NIMBLE model,

\code{observedStates[1:T] ~ dHMM(initStates[1:S], observationProbs[1:O, 1:S],
transitionProbs[1:S, 1:S], T)}

declares that the \code{observedStates[1:T]} vector follows a hidden Markov model distribution
with parameters as indicated, assuming all the parameters have been declared elsewhere in the model.  In
this case, \code{S} is the number of system states and \code{O} is the number of observation classes, and
\code{T} is the number of observation occasions.

If the observation probabilities are time-dependent, one would use:

\code{observedStates[1:T] ~ dHMMo(initStates[1:S], observationProbs[1:O, 1:S, 1:T],
transitionProbs[1:S, 1:S], T)}
}
\examples{
\dontrun{
# Set up constants and initial values for defining the model
len <- 5 # length of dataset
dat <- c(1,2,1,1,2) # A vector of observations
init <- c(0.4, 0.2, 0.4) # A vector of initial state probabilities
probObs <- t(array( # A matrix of observation probabilities
       c(1, 0.2, 1,
         0, 0.8, 0), c(3, 2)))
probTrans <- t(array( # A matrix of transition probabilities
        c(0.6, 0.3, 0.1,
          0, 0.7, 0.3,
          0, 0, 1), c(3,3)))

# Define code for a nimbleModel
 nc <- nimbleCode({
   x[1:5] ~ dHMM(init[1:3], probObs = probObs[1:2,1:3],
                 probTrans = probTrans[1:3, 1:3], len = 5)

   for (i in 1:3) {
     init[i] ~ dunif(0,1)

     for (j in 1:3) {
       probTrans[i,j] ~ dunif(0,1)
     }

     probTrans[1,i] ~ dunif(0,1)
     probTrans[2,i] <- 1 - probTrans[1,i]
   }
 })

# Build the model
HMM_model <- nimbleModel(nc,
                         data = list(x = dat),
                         inits = list(init = init,
                                      probObs = probObs,
                                      probTrans = probTrans)))
# Calculate log probability of data from the model
HMM_model$calculate()
# Use the model for a variety of other purposes...
}
}
\references{
D. Turek, P. de Valpine and C. J. Paciorek. 2016. Efficient Markov chain Monte
Carlo sampling for hierarchical hidden Markov models. Environmental and Ecological Statistics
23:549–564. DOI 10.1007/s10651-016-0353-z
}
\seealso{
For dynamic hidden Markov models with time-dependent transitions, see \link{dDHMM} and \link{dDHMMo}.
For simple capture-recapture, see \link{dCJS}.
}
\author{
Ben Goldstein, Perry de Valpine, and Daniel Turek
}
