% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dOcc.R
\name{dOcc}
\alias{dOcc}
\alias{dOcc_ss}
\alias{dOcc_sv}
\alias{dOcc_vs}
\alias{dOcc_vv}
\alias{rOcc_ss}
\alias{rOcc_sv}
\alias{rOcc_vs}
\alias{rOcc_vv}
\title{Occupancy distribution for use in NIMBLE models}
\usage{
dOcc_ss(x, probOcc, probDetect, len = 0, log = 0)

dOcc_sv(x, probOcc, probDetect, len = 0, log = 0)

dOcc_vs(x, probOcc, probDetect, len = 0, log = 0)

dOcc_vv(x, probOcc, probDetect, len = 0, log = 0)

rOcc_ss(n, probOcc, probDetect, len = 0)

rOcc_sv(n, probOcc, probDetect, len = 0)

rOcc_vs(n, probOcc, probDetect, len = 0)

rOcc_vv(n, probOcc, probDetect, len = 0)
}
\arguments{
\item{x}{detection/non-detection vector of 0s (not detected) and 1s (detected).}

\item{probOcc}{occupancy probability (scalar).}

\item{probDetect}{detection probability (scalar for \code{dOcc_s}, vector for \code{dOcc_v}).}

\item{log}{TRUE (return log probability) or FALSE (return probability)}

\item{l}{length of detection/non-detection vector (ignored for "d" functions, needed for "r" functions).}
}
\description{
\code{dOcc_s} and \code{dOcc_v} provide occupancy model distributions for NIMBLE models.
The 's' version uses time-independent (scalar) detection probability while the 'v' version
uses time-dependent (vector) detection probability.
Compared to writing NIMBLE models with a discrete latent state for true occupancy status and
a separate scalar datum for each observation,
use of these distributions allows
one to directly sum over the discrete latent state and calculate the probability of
all observations from one site jointly.
}
\details{
These nimbleFunctions provide distributions that can be used in code (via \link{nimbleCode})
for \link{nimbleModel}.

These are written in the format of user-defined distributions to extend NIMBLE's
use of the BUGS model language.  More information about writing user-defined distributions can be found
in the NIMBLE User Manual at \code{https://r-nimble.org}.

The first argument to a "d" function is always named \code{x} and is given on the
left-hand side of a (stochastic) model declaration in the BUGS model language (used by NIMBLE).
When using these distributions in a NIMBLE model, the user
should not provide the \code{log} argument.  (It is always set to \code{TRUE} when used
in a NIMBLE model.)

For example, in a NIMBLE model,

\code{detections[1:T] ~ dOcc_s(occupancyProbability, detectionProbability)}

declares that the \code{detections[1:T]} vector follows an occupancy model distribution
with parameters as indicated, assuming all the parameters have been declared elsewhere in the model.

If the detection probabilities are time-dependent, one would use:

\code{detections[1:T] ~ dOcc_s(occupancyProbability, detectionProbability[1:T])}
}
\seealso{
For dynamic occupancy models, see \link{dDynOcc}.
}
\author{
Ben Goldstein and Perry de Valpine
}
