% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dOcc.R
\name{dOcc_s}
\alias{dOcc_s}
\alias{dOcc}
\alias{rOcc}
\alias{rOcc_s}
\alias{dOcc_v}
\alias{rOcc_v}
\title{Occupancy distribution for use in NIMBLE models}
\usage{
dOcc_s(x, probOcc, probDetect, log = 0)
}
\arguments{
\item{x}{detection/non-detection vector of 0s (not detected) and 1s (detected).}

\item{probOcc}{occupancy probability}

\item{probDetect}{detection probability}

\item{log}{TRUE (return log probability) or FALSE (return probability)}

\item{l}{length of detection/non-detection history (needed for rOcc)}
}
\description{
\code{dOcc_s} and \code{dOcc_v} provide occupancy model distributions for NIMBLE models.
The 's' version uses time-independent (scalar) detection probability while the 'v' version
uses time-dependent (vector) detection probability.
Compared to writing NIMBLE models with a discrete latent state for true occupancy status and
a separate scalar datum for each observation,
use of these distributions allows
one to directly sum over the discrete latent state and calculate the probability of
all observations from one site jointly.
}
\details{
To be filled in.
}
