% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{multispeciesOccupancy}
\alias{multispeciesOccupancy}
\title{Macro to fit single-season, multi-species occupancy model}
\arguments{
\item{stateformula}{An R formula for the occupancy/state model, possibly with the 
parameters followed by brackets containing indices.}

\item{detformula}{An R formula for the detection model, possibly with the 
parameters followed by brackets containing indices.}

\item{statePrefix}{All state model coefficient names will begin with this prefix.
default is state_ (so x becomes state_x, etc.)}

\item{detPrefix}{All detection model coefficient names will begin with this prefix.
default is det_ (so x becomes det_x, etc.)}

\item{statePriors}{Prior specifications for state model parameters, should be 
generated with nimbleMacros::setPriors()}

\item{detPriors}{Prior specifications for det model parameters, should be generated 
with nimbleMacros::setPriors()}

\item{marginalized}{Logical. If TRUE, fit the marginalized model using the
dOcc_v nimbleFunction}

\item{centerVar}{Grouping covariate to 'center' on in parameterization. By
default all random effects have mean 0 as with lme4.}

\item{speciesID}{Name to use for species ID index vector}
}
\description{
Generates nimble code for a single-season, multi-species occupancy model.
Covariates can be specified for both the detection and occupancy (state)
submodels using R formulas.
}
\examples{
nimbleOptions(enableModelMacros = TRUE)
}
\author{
Ken Kellner
}
