% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleOccu.R
\name{nimbleOccu}
\alias{nimbleOccu}
\title{Fit occupancy models with nimble}
\usage{
nimbleOccu(
  stateformula,
  detformula,
  y,
  siteCovs = NULL,
  obsCovs = NULL,
  speciesCovs = NULL,
  statePriors = setPriors(intercept = "dunif(-10, 10)", coefficient = "dlogis(0, 1)", sd
    = "dhalfflat()"),
  detPriors = setPriors(intercept = "dunif(-10, 10)", coefficient = "dlogis(0, 1)", sd =
    "dhalfflat()"),
  marginalized = FALSE,
  returnModel = FALSE,
  sampler = c("default", "polyagamma", "RW_block", "barker", "hmc"),
  savePsi = TRUE,
  saveP = FALSE,
  samplerControl = list(),
  buildDerivs = FALSE,
  samplesOnly = FALSE,
  ...
)
}
\arguments{
\item{stateformula}{An R formula for the occupancy/state model, possibly with the 
parameters followed by brackets containing indices.}

\item{detformula}{An R formula for the detection model, possibly with the 
parameters followed by brackets containing indices.}

\item{y}{The detection/non-detection data. For single-species models this 
should be an MxJ matrix; for multi-species models it should be an MxJxS
array, where M is sites, J is occasions, and S is species.}

\item{siteCovs}{A named list of site-level covariates. Each should be a vector of length M.}

\item{obsCovs}{A named list of observation-level covariates. Each should be 
a matrix of dimensions MxJ.}

\item{speciesCovs}{A named list of species-level covariates. Each should be
a vector of length S. Only used in multi-species model.}

\item{statePriors}{Prior specifications for state model parameters, should be 
generated with nimbleMacros::setPriors()}

\item{detPriors}{Prior specifications for det model parameters, should be generated 
with nimbleMacros::setPriors()}

\item{marginalized}{Logical. If TRUE, fit the marginalized model using the
dOcc_v nimbleFunction}

\item{returnModel}{Logical. If TRUE, the nimble model object is returned.}

\item{sampler}{MCMC sampler to use. Currently only the nimble default sampler works.}

\item{savePsi}{Save site-level occupancy probabilities in the output?}

\item{saveP}{Save site x occasion level detection probabilities in the output?}

\item{samplerControl}{An optional list of control arguments to sampler functions 
to be passed `configureMCMC`. In addition, one can provide the element
\code{noncentered = TRUE} to add NIMBLE's "noncentered" sampler to jointly sample
each hyperparameter with its dependent random effects.}

\item{buildDerivs}{Build derivative capabilities for the model?}

\item{samplesOnly}{Save only the matrix of samples? If \code{FALSE}, the model is also saved.}

\item{...}{Arguments passed to \code{runMCMC} such as \code{nchains}.}
}
\value{
Either MCMC samples, or a nimble model object if \code{returnModel = TRUE}.
}
\description{
Fit single or multi-species occupancy models with nimble. The state
and detection models are defined with R formulas and data are supplied to
the y, siteCovs, obsCovs, and speciesCovs arguments.
You may either fit the model or return the model object for further processing..
}
\author{
Ken Kellner
}
