% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{occupancy}
\alias{occupancy}
\title{Macro to fit single-season, single-species occupancy model}
\arguments{
\item{stateformula}{An R formula for the occupancy/state model, possibly with the 
parameters followed by brackets containing indices.}

\item{detformula}{An R formula for the detection model, possibly with the 
parameters followed by brackets containing indices.}

\item{statePrefix}{All state model coefficient names will begin with this prefix.
default is state_ (so x becomes state_x, etc.)}

\item{detPrefix}{All detection model coefficient names will begin with this prefix.
default is det_ (so x becomes det_x, etc.)}

\item{statePriors}{Prior specifications for state model parameters, should be 
generated with nimbleMacros::setPriors()}

\item{detPriors}{Prior specifications for det model parameters, should be generated 
with nimbleMacros::setPriors()}

\item{marginalized}{Logical. If TRUE, fit the marginalized model using the
dOcc_v nimbleFunction}

\item{centerVar}{Grouping covariate to 'center' on in parameterization. By
default all random effects have mean 0 as with lme4.}
}
\description{
Generates nimble code for a single-season, single-species occupancy model.
Covariates can be specified for both the detection and occupancy (state)
submodels using R formulas.
}
\details{
This macro should be placed on the right-hand side of an assignment
 (it doesn't matter if you use <- or ~). The left-hand side must be a
 matrix of detection-nondetection data with dimensions M x J, where
 M is sites and J is occasions. This matrix may contain
 missing values (e.g. if you have different numbers of occasions by site). However,
 to use the marginalized form of the model, the missing values for site M
 must come at the end of the row of occasion data 1:J. For example
 if you have J = 5 but the first site was only sampled 3 times, the 
 corresponding row of the matrix should be [1, 1, 1, NA, NA] rather than
 for example [1, NA, 1, 1, NA].
}
\examples{
nimbleOptions(enableModelMacros = TRUE)
y <- matrix(rbinom(10, 1, 0.5), 5, 2)
x <- rnorm(5)
const <- list(y=y, x=x, M=nrow(y), J = ncol(y))

code <- nimbleCode({
  y[1:M, 1:J] ~ occupancy(~x[1:M], ~1)
})

mod <- nimbleModel(code, constants=const)
mod$getCode()
}
\author{
Ken Kellner
}
