% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCJS.R
\name{dCJS}
\alias{dCJS}
\alias{dCJSss}
\alias{dCJSsv}
\alias{dCJSvs}
\alias{dCJSvv}
\alias{rCJSss}
\alias{rCJSsv}
\alias{rCJSvs}
\alias{rCJSvv}
\title{Cormack-Jolly-Seber distribution for use in NIMBLE models}
\usage{
dCJSss(x, probSurvive, probCapture, len = 0, log = 0)

dCJSsv(x, probSurvive, probCapture, len = 0, log = 0)

dCJSvs(x, probSurvive, probCapture, len = 0, log = 0)

dCJSvv(x, probSurvive, probCapture, len = 0, log = 0)

rCJSss(n, probSurvive, probCapture, len = 0)

rCJSsv(n, probSurvive, probCapture, len = 0)

rCJSvs(n, probSurvive, probCapture, len = 0)

rCJSvv(n, probSurvive, probCapture, len = 0)
}
\arguments{
\item{x}{capture-history vector of 0s (not captured) and 1s (captured). Do not include the initial capture, which is assumed.}

\item{probSurvive}{survival probability, either a scalar (for dCJSs*) or a vector (for dCJSv*).}

\item{probCapture}{capture probability, either a scalar (for dCJS*s) or a vector (for dCJS*v).}

\item{len}{length of capture history (needed for rCJSxx).}

\item{log}{TRUE or 1 to return log probability. FALSE or 0 to return probability.}

\item{n}{}
}
\description{
\code{dCJS**} functions provide a basic distribution for capture history vectors based
 on survival and capture probabilities. The different aliases are for scalar ("s", time-independent)
 versus vector ("v", time-dependent) survival and capture probabilities, in that order.
}
\details{
These nimbleFunctions provide distributions that can be used in code (via \code{nimbleCode}) for \link{nimbleModel}.

These are written in the format of user-defined distributions to extend NIMBLE's
use of the BUGS model language.  More information about writing user-defined distributions can be found
in the NIMBLE User Manual at \code{https://r-nimble.org}.

The first argument to a "d" function is always named \code{x} and is given on the
left-hand side of a (stochastic) model declaration in the BUGS model language (used by NIMBLE).
When using these distributions in a NIMBLE model, the user
should not provide the \code{log} argument.  (It is always set to \code{TRUE} when used
in a NIMBLE model.)

For example,

\code{captures[1:T] ~ dCSJss(survive, capture, T)}

declares a vector node, \code{captures[1:T]}, that follows a capture-recapture distribution
with scalar survival probability \code{survive} and scalar capture probability \code{capture}
(assuming \code{survive} and \code{capture} are defined elsewhere in the model).
If time-dependent survival and capture probabilities are needed, use

\code{captures[1:T] ~ dCSJss(survive[1:T], capture[1:T], T)}.

In fact, the \code{len} argument (\code{T} in these examples) will be ignored during model
probability calculations.  It will be used only for simulations.

\emph{Ignored elements of \code{survive} and \code{capture} when they are vectors:} Note that the last element of \code{survive},
if present, will be ignored.  Also the first element of \code{capture} will be ignored, since the first
element of the capture history is assumed to be 1, indicated first capture.
}
\references{
D. Turek, P. de Valpine and C. J. Paciorek. 2016. Efficient Markov chain Monte
Carlo sampling for hierarchical hidden Markov models. Environmental and Ecological Statistics
23:549–564. DOI 10.1007/s10651-016-0353-z
}
\seealso{
For multi-state or multi-event capture-recapture models, see \link{dHMM} or \link{dDHMM}.
}
\author{
Ben Goldstein, Perry de Valpine, and Daniel Turek
}
