% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dNmixture.R
\name{dNmixture}
\alias{dNmixture}
\alias{dNmixture_v}
\alias{dNmixture_s}
\alias{rNmixture_v}
\alias{rNmixture_s}
\title{N-mixture distribution for use in NIMBLE models}
\usage{
dNmixture_v(x, lambda, prob, minN, maxN, len, log = 0)

dNmixture_s(x, lambda, prob, minN, maxN, len, log = 0)

rNmixture_v(n, lambda, prob, minN, maxN, len)

rNmixture_s(n, lambda, prob, minN, maxN, len)
}
\arguments{
\item{x}{count observation data, 0 and positive integers}

\item{lambda}{expected value of the Poisson distribution of true abundance}

\item{prob}{observation probability for each X}

\item{minN}{the minimum true abundance to sum over. Set to -1 for distribution
to select based on lambda. Ignored if dynamicMinMax = TRUE}

\item{maxN}{the maximum true abundance to sum over. Set to -1 for distribution
to select based on lambda. Ignored if dynamicMinMax = TRUE}

\item{len}{The length of the x vector. Needed for simulation in \code{rNmixture_*}.}

\item{n}{number of random draws, each returning a vector of length
\code{len}. Currently only \code{n = 1} is supported, but the
argument exists for standardization of "\code{r}" functions.}

\item{notZero}{0 if datum is structural zero, 1 otherwise. Allows for
zero-inflated Poisson models}

\item{dynamicMinMax}{0 to use input minN and maxN, 1 to ignore
provided minN and maxN and algorithmically select reasonable bounds for N.}
}
\description{
\code{dNmixture} provides dynamic occupancy model distributions for NIMBLE models.
}
\details{
N-mixture models
model abundance at a series of sites over many replicate observations. The likelihood of an observation in site s
at time t depends on the

Compared to writing NIMBLE models with a discrete latent state for true occupancy status and
a separate scalar datum for each observation,
use of these distributions allows
one to directly sum over the discrete latent state and calculate the probability of
all observations from one site jointly.
}
\examples{
\dontrun{}
}
\seealso{
For regular occupancy models, see documentation for dOcc.
}
\author{
Lauren Ponisio, Ben Goldstein, Perry de Valpine

These are written in the format of user-defined distributions to extend NIMBLE's
use of the BUGS model language. More information about writing user-defined distributions can be found
in the NIMBLE User Manual at \code{https://r-nimble.org}.

The first argument to a "d" function is always named \code{x} and is given on the
left-hand side of a (stochastic) model declaration in the BUGS model language (used by NIMBLE).
When using these distributions in a NIMBLE model, the user
should not provide the \code{log} argument. (It is always set to \code{TRUE} when used
in a NIMBLE model.)

For example, in a NIMBLE model,

\code{detections[1:T] ~ dNmixture(lambda = lambda, prob = p[1:T], notZero = 1, minN = -1, maxN = -1)}

declares that the \code{detections[1:T]} vector follows an N-mixture model distribution
with parameters as indicated, assuming all the parameters have been declared elsewhere in the model.
}
