% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dNmixture_MNB.R, R/dNmixture_MP.R
\name{dNmixture_M}
\alias{dNmixture_M}
\alias{dNmixture_MNB_s}
\alias{dNmixture_MNB_v}
\alias{rNmixture_MNB_s}
\alias{rNmixture_MNB_v}
\alias{dNmixture_MP_s}
\alias{dNmixture_MP_v}
\alias{rNmixture_MP_s}
\alias{rNmixture_MP_v}
\title{Multinomial N-mixture distribution for use in \code{nimble} models}
\usage{
dNmixture_MNB_s(x, lambda, p, theta, J, log = 0)

rNmixture_MNB_s(n, lambda, p, theta, J)

dNmixture_MNB_v(x, lambda, p, theta, J, log = 0)

rNmixture_MNB_v(n, lambda, p, theta, J)

dNmixture_MP_s(x, lambda, p, J, log = 0)

rNmixture_MP_s(n, lambda, p, J)

dNmixture_MP_v(x, lambda, p, J, log = 0)

rNmixture_MP_v(n, lambda, p, J)
}
\arguments{
\item{x}{vector of integer counts from a series of sampling occasions.}

\item{lambda}{expected value of the (negative binomial or Poisson)
distribution of true abundance.}

\item{p}{detection probability (scalar for \code{dNmixture_M\*_s},
vector for \code{dNmixture_M\*_v}). If a vector, p[i] is the probability
that an individual is observed in observation category i and no other
categories. \code{p} must sum to less than or equal to 1, although this
is not checked by the function for computational efficiency. The
probability of an individual going unobserved is \code{1 - sum(p)}.}

\item{theta}{overdispersion parameter required for negative binomial
(*NB) models. theta is parameterized such that variance of
the negative binomial variable x is \code{lambda^2 * theta + lambda}}

\item{J}{integer number of searches.}

\item{log}{\code{TRUE} or \code{1} to return log probability. \code{FALSE} or
\code{0} to return probability.}

\item{n}{number of random draws, each returning a vector of length
\code{len}. Currently only \code{n = 1} is supported, but the argument
exists for standardization of "\code{r}" functions.}
}
\value{
For \code{dNmixture_s} and \code{dNmixture_v}: the probability (or likelihood) or log probability of observation vector \code{x}. \cr
        For \code{rNmixture_s} and \code{rNmixture_v}: a simulated detection history, \code{x}.
}
\description{
\code{dNmixture_MNB_s} and \code{dNmixture_MNB_v} provide Multinomial Negative Binomial (MNB) mixture distributions of abundance ("N-mixture") for use in \code{nimble} models.
}
\details{
These nimbleFunctions provide distributions that can be used directly in R or in \code{nimble} hierarchical models (via \code{\link[nimble]{nimbleCode}} and \code{\link[nimble]{nimbleModel}}). \cr
         The distributions are implemented in closed-form following Haines (2020), which avoids infinite sum-based calculations.
}
\examples{

# Set up variables and parameters
#  J:  number of visits
#  lambda: mean abundance
#  theta:  scale parameter on abundance distribution
#  p:  search-specific detection probabilities

 J <- 10
 lambda <- 5
 theta  <- 2
 p  <- runif(J, 0.4, 0.7)

 lambdat <- log(mu)
 thetat  <- log(theta)

# Generate data

  yv <- rNmixture_MNB_v(n = 1, lambda = lambda, p = p, theta = theta, J = J)

# Write the model code

  nc <- nimbleCode({

    lambdat ~  dnorm(0, 1/2)
    lambda  <- exp(lambdat)

    thetat ~  dnorm(0, 0.1)
    theta  <- exp(thetat)

    for (j in 1:J) {
      p[j] ~ dunif(0, 1)
    }

    x[1:J] ~ dNmixture_MNB_v(lambda = lambda, p = p[1:J], theta = theta, J = J)

  })

# Construct the model object

  nmix <- nimbleModel(nc,
                      constants = list(J = J),
                      data = list(x = yv),
                      inits = list(lambdat = lambdat,
                                   thetat  = thetat,
                                   p   = p))

# Calculate log probability of data from the model

  nmix$calculate()


}
\references{
L. M. Haines. 2020. Multinomial N-mixture models for removal sampling. Biometrics 76:540-548. DOI 10.1111/biom.13147
}
\seealso{
For binomial N-mixture models, see \code{\link{dNmixture}}. \cr
         For occupancy models dealing with detection/nondetection data, see \code{\link{dOcc}}. \cr
         For dynamic occupancy, see \code{\link{dDynOcc}}.
}
\author{
Juniper Simonis, Ben Goldstein
}
