% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dOcc.R
\name{dOcc}
\alias{dOcc}
\alias{dOcc_s}
\alias{dOcc_v}
\alias{rOcc_s}
\alias{rOcc_v}
\title{Occupancy distribution for use in NIMBLE models}
\usage{
dOcc_s(x, probOcc, probDetect, len = 0, log = 0)

dOcc_v(x, probOcc, probDetect, len = 0, log = 0)

rOcc_s(n, probOcc, probDetect, len = 0)

rOcc_v(n, probOcc, probDetect, len = 0)
}
\arguments{
\item{x}{detection/non-detection vector of 0s (not detected) and 1s (detected).}

\item{probOcc}{occupancy probability (scalar for \code{dOcc_s*}, vector for \code{dOcc_v*}).}

\item{probDetect}{detection probability (scalar for \code{dOcc_*s}, vector for \code{dOcc_*v}).}

\item{len}{length of detection/non-detection vector (ignored for "d" functions, needed for "r" functions).}

\item{log}{TRUE (return log probability) or FALSE (return probability)}

\item{n}{length of random sequence}
}
\description{
\code{dOcc_**} provides occupancy model distributions for NIMBLE models.
}
\details{
Likelihood of observation x[t] depends on an occupancy probability \code{probOcc[t]} and
detection probability \code{probDetect[t]}.
The pair of letters following the 'dOcc_' indicates whether the occupancy probability
and detection probability are scalar (s) or vector (v). For example, dOcc_sc takes scalar
occupancy probability with a vector of detection probabilities.

Compared to writing NIMBLE models with a discrete latent state for true occupancy status and
a separate scalar datum for each observation,
use of these distributions allows
one to directly sum over the discrete latent state and calculate the probability of
all observations from one site jointly.

These nimbleFunctions provide distributions that can be used in code (via \link{nimbleCode})
for \link{nimbleModel}.

These are written in the format of user-defined distributions to extend NIMBLE's
use of the BUGS model language. More information about writing user-defined distributions can be found
in the NIMBLE User Manual at \code{https://r-nimble.org}.

The first argument to a "d" function is always named \code{x} and is given on the
left-hand side of a (stochastic) model declaration in the BUGS model language (used by NIMBLE).
When using these distributions in a NIMBLE model, the user
should not provide the \code{log} argument. (It is always set to \code{TRUE} when used
in a NIMBLE model.)

For example, in a NIMBLE model,

\code{detections[1:T] ~ dOcc_ss(occupancyProbability, detectionProbability)}

declares that the \code{detections[1:T]} vector follows an occupancy model distribution
with parameters as indicated, assuming all the parameters have been declared elsewhere in the model.

If the detection probabilities are time-dependent, one would use:

\code{detections[1:T] ~ dOcc_sv(occupancyProbability, detectionProbability[1:T])}
}
\examples{
\dontrun{
# Set up constants and initial values for defining the model
dat <- c(1,1,0,0) # A vector of observations
probOcc <- 0.6
probDetect <- 0.4


# Define code for a nimbleModel
nc <- nimbleCode({
  x[1:4] ~ dOcc_s(probOcc, probDetect, len = 4)
  probOcc ~ dunif(0,1)
  probDetect ~ dunif(0,1)
})

# Build the model, providing data and initial values
Occ_model <- nimbleModel(nc, data = list(x = dat),
                         inits = list(probOcc = probOcc,
                                      probDetect = probDetect))

# Calculate log probability of data from the model
Occ_model$calculate()
# Use the model for a variety of other purposes...
}
}
\seealso{
For dynamic occupancy models, see documentation for \link{dDynOcc}.
}
\author{
Ben Goldstein and Perry de Valpine
}
