% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{MULTISPECIESOCCUPANCY}
\alias{MULTISPECIESOCCUPANCY}
\title{Macro to fit single-season, multi-species occupancy model}
\arguments{
\item{stateformula}{An R formula for the occupancy/state model, possibly with the 
parameters followed by brackets containing indices.}

\item{detformula}{An R formula for the detection model, possibly with the 
parameters followed by brackets containing indices.}

\item{statePrefix}{All state model coefficient names will begin with this prefix.
default is state_ (so x becomes state_x, etc.)}

\item{detPrefix}{All detection model coefficient names will begin with this prefix.
default is det_ (so x becomes det_x, etc.)}

\item{statePriors}{Prior specifications for state model parameters, should be 
generated with nimbleMacros::setPriors()}

\item{detPriors}{Prior specifications for det model parameters, should be generated 
with nimbleMacros::setPriors()}

\item{marginalized}{Logical. If TRUE, fit the marginalized model using the
dOcc_v nimbleFunction}

\item{speciesID}{Name to use for species ID index vector}
}
\description{
Generates nimble code for a single-season, multi-species occupancy model,
in which all intercepts and slopes are assumed to be random effects by species.
Covariates can be specified for both the detection and occupancy (state)
submodels using R formulas.
}
\details{
This macro should be placed on the right-hand side of an assignment
 (it doesn't matter if you use <- or ~). The left-hand side must be an
 array of detection-nondetection data with dimensions M x J x S, where
 M is sites, J is occasions, and S is species.  This array may contain
 missing values (e.g. if you have different numbers of occasions by site).
}
\examples{
nimbleOptions(enableModelMacros = TRUE)
M <- 100 # sites
J <- 5   # occasions
S <- 10  # species
y <- array(rbinom(M*J*S, 1, 0.5), c(M,J,S))
x <- rnorm(M) # site covariate

const <- list(M=M, J=J, S=S, y=y, x=x)

code <- nimbleCode({
  y[1:M, 1:J, 1:S] ~ MULTISPECIESOCCUPANCY(~x[1:M], ~1)
})

mod <- nimbleModel(code, constants=const)
mod$getCode()
}
\author{
Ken Kellner
}
