% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nimNmixPois_logFac}
\alias{nimNmixPois_logFac}
\title{Helper function for fast N-mixture calculation}
\usage{
nimNmixPois_logFac(numN, ff)
}
\arguments{
\item{numN}{first argument for helper function nimNmixPois_logFac,
representing the number of indices in the truncated sum for the N-mixture.}

\item{ff}{second argument for helper function nimNmixPois_logFac, a derived
vector of units calculated partway through the fast N-mixture algorithm.}
}
\description{
Helper function for fast N-mixture calculation
}
