% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{occupancy}
\alias{occupancy}
\title{Macro to fit single-season, single-species occupancy model}
\arguments{
\item{stateformula}{An R formula for the occupancy/state model, possibly with the 
parameters followed by brackets containing indices.}

\item{detformula}{An R formula for the detection model, possibly with the 
parameters followed by brackets containing indices.}

\item{statePrefix}{All state model coefficient names will begin with this prefix.
default is state_ (so x becomes state_x, etc.)}

\item{detPrefix}{All detection model coefficient names will begin with this prefix.
default is det_ (so x becomes det_x, etc.)}

\item{statePriors}{Prior specifications for state model parameters, should be generated with nimbleMacros::setPrior()}

\item{detPriors}{Prior specifications for det model parameters, should be generated with nimbleMacros::setPrior()}
}
\description{
Generates nimble code for a single-season, single-species occupancy model.
Covariates can be specified for both the detection and occupancy (state)
submodels using R formulas.
}
\examples{
code <- nimbleCode({
  mu[1:10] <- linPred(~x + x2)
})

mod <- nimbleModel(code)
mod$getCode()
}
\author{
Ken Kellner
}
